+++ android/app/src/main/java/pro/congviec/app/MainActivity.java
@@ -1,0 +1,12 @@
+package pro.congviec.app;
+
+import android.os.Bundle;
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(BackupFilePlugin.class);
+        super.onCreate(savedInstanceState);
+    }
+}